chore: prepare Release 52 for release testing
