+++ meteor/server/migration/migrations.ts
@@ -37,6 +37,9 @@ addSteps1_50_0()
 import { addSteps as addSteps1_51_0 } from './1_51_0'
 addSteps1_51_0()
 
+import { addSteps as addSteps1_52_0 } from './1_52_0'
+addSteps1_52_0()
+
 // Migrations for the in-development release:
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
